restructuring of build system

+++ vistle/core/object.cpp
@@ -29,7 +29,7 @@ Shm & Shm::instance(const int moduleID, const int rank,
                     message::MessageQueue * mq) {
 
    if (!singleton)
-      singleton = new Shm(moduleID, rank, LONG_MAX, mq);
+      singleton = new Shm(moduleID, rank, INT_MAX, mq);
 
    return *singleton;
 }
@@ -37,7 +37,7 @@ Shm & Shm::instance(const int moduleID, const int rank,
 Shm & Shm::instance() {
 
    if (!singleton)
-      singleton = new Shm(-1, -1, ULONG_MAX, NULL);//34359738368); // 32GB
+      singleton = new Shm(-1, -1, INT_MAX, NULL);//34359738368); // 32GB
 
    return *singleton;
 }
